feat(execute,project-loader): use indexing workaround
* Move forceIndexing flag from ModelCheckArgs and GenerateArgs to Args.
* Represent the flag in ProjectLoader and ProjectLoaderBuilder.
* Call the indexing workaround in ProjectLoader after opening a project.
