+++ gradle.properties
@@ -1,7 +1,7 @@
 org.gradle.parallel=true
 
-version.backend=1.24.1
-version.project-loader=3.0.2
+version.backend=1.24.2
+version.project-loader=4.0.0
 
 # A comma-separated list of MPS releases or prereleases to test against.
 supportedMpsVersions=2022.3.3,2023.2.2,2024.1.2,2024.3.1,251.23774.10091
