+++ project-loader/api/project-loader.api
@@ -20,6 +20,7 @@ public final class de/itemis/mps/gradle/logging/LoggingFactoryKt {
 public class de/itemis/mps/gradle/project/loader/Args : de/itemis/mps/gradle/project/loader/EnvironmentArgs {
 	public fun <init> (Lcom/xenomachina/argparser/ArgParser;)V
 	public fun configureProjectLoader (Lde/itemis/mps/gradle/project/loader/ProjectLoader$Builder;)V
+	public final fun getForceIndexing ()Ljava/lang/Boolean;
 	public final fun getProject ()Ljava/io/File;
 }
 
@@ -62,11 +63,6 @@ public final class de/itemis/mps/gradle/project/loader/EnvironmentKind : java/la
 	public static final field MPS Lde/itemis/mps/gradle/project/loader/EnvironmentKind;
 	public static fun valueOf (Ljava/lang/String;)Lde/itemis/mps/gradle/project/loader/EnvironmentKind;
 	public static fun values ()[Lde/itemis/mps/gradle/project/loader/EnvironmentKind;
-}
-
-public final class de/itemis/mps/gradle/project/loader/IndexingWorkaroundKt {
-	public static final fun forceIndexing (Ljetbrains/mps/project/MPSProject;Lcom/intellij/openapi/util/BuildNumber;)V
-	public static final fun hasIndexingBug (Lcom/intellij/openapi/util/BuildNumber;)Z
 }
 
 public final class de/itemis/mps/gradle/project/loader/LoaderKt {
@@ -122,7 +118,7 @@ public final class de/itemis/mps/gradle/project/loader/Plugin {
 
 public final class de/itemis/mps/gradle/project/loader/ProjectLoader {
 	public static final field Companion Lde/itemis/mps/gradle/project/loader/ProjectLoader$Companion;
-	public synthetic fun <init> (Ljetbrains/mps/tool/environment/EnvironmentConfig;Lde/itemis/mps/gradle/project/loader/EnvironmentKind;Ljava/io/File;Ljava/lang/String;Lde/itemis/mps/gradle/logging/LogLevel;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public synthetic fun <init> (Ljetbrains/mps/tool/environment/EnvironmentConfig;Lde/itemis/mps/gradle/project/loader/EnvironmentKind;Ljava/io/File;Ljava/lang/String;Lde/itemis/mps/gradle/logging/LogLevel;Ljava/lang/Boolean;Lkotlin/jvm/internal/DefaultConstructorMarker;)V
 	public final fun execute (Lkotlin/jvm/functions/Function1;)Ljava/lang/Object;
 	public final fun executeWithProject (Ljava/io/File;Lkotlin/jvm/functions/Function2;)Ljava/lang/Object;
 }
@@ -134,9 +130,11 @@ public final class de/itemis/mps/gradle/project/loader/ProjectLoader$Builder {
 	public final fun getBuildNumber ()Ljava/lang/String;
 	public final fun getEnvironmentConfigBuilder ()Lde/itemis/mps/gradle/project/loader/EnvironmentConfigBuilder;
 	public final fun getEnvironmentKind ()Lde/itemis/mps/gradle/project/loader/EnvironmentKind;
+	public final fun getForceIndexing ()Ljava/lang/Boolean;
 	public final fun getLogLevel ()Lde/itemis/mps/gradle/logging/LogLevel;
 	public final fun setBuildNumber (Ljava/lang/String;)V
 	public final fun setEnvironmentKind (Lde/itemis/mps/gradle/project/loader/EnvironmentKind;)V
+	public final fun setForceIndexing (Ljava/lang/Boolean;)V
 	public final fun setLogLevel (Lde/itemis/mps/gradle/logging/LogLevel;)V
 }
